// PatientPortal/src/app/patientChart/models/baseVitalSigns.ts
export class BaseVitalSigns {
    id: string;
    patientId: string;
    dominantHand: string;
    weight: number;
    height: number;
    leftBicep: number;
    rightBicep: number;
    leftForearm: number;
    rightForearm: number;
    leftThigh: number;
    rightThigh: number;
    leftCalf: number;
    rightCalf: number;
    oxygenUse: string;
    oxygenAmount: number;
    headCircumference: number;
}